divide penbox from element_commands
